Split up JsonRDD2 into multiple objects
